Final Commit

+++ app/src/main/java/com/example/mygame/Player.java
@@ -5,7 +5,7 @@ import androidx.core.content.ContextCompat;
 
 public class Player extends Circle  {
     public float EatenFood = 0;
-    public static  double MAX_SPEED = 10;
+    public static double MAX_SPEED = 10;
     public double mass;
     private final Joystick joystick;
 
@@ -26,7 +26,7 @@ public class Player extends Circle  {
         Collision();
         }
 
-    public  boolean Can_SEE_FOOD(int pX,int pY,int wp,int hp){
+    public  boolean Can_SEE_FOOD(float pX,float pY,int wp,int hp){
         return (positionX+wp>pX && positionX-wp<pX && positionY+hp>pY && positionY-hp<pY);
     }
     }
